Splits states and contracts from other source files so that only shared structures have to be sent over the wire.
