+++ .gitignore
@@ -56,6 +56,8 @@ lib/dokka.jar
 
 # IntelliJ
 /out/
+states-and-contracts/out/
+other-sources/out/
 
 # mpeltonen/sbt-idea plugin
 .idea_modules/
