+++ other-sources/build.gradle
@@ -64,20 +64,17 @@ dependencies {
     cordaCompile "net.corda:corda-core:$corda_release_version"
     cordaCompile "net.corda:corda-finance:$corda_release_version"
     cordaCompile "net.corda:corda-jackson:$corda_release_version"
-    cordaCompile "net.corda:corda-rpc:$corda_release_version"
     cordaCompile "net.corda:corda-node-api:$corda_release_version"
+    cordaCompile "net.corda:corda-rpc:$corda_release_version"
     cordaCompile "net.corda:corda-webserver-impl:$corda_release_version"
     cordaRuntime "net.corda:corda:$corda_release_version"
     cordaRuntime "net.corda:corda-webserver:$corda_release_version"
-
     testCompile "net.corda:corda-node-driver:$corda_release_version"
 
     // CorDapp dependencies
     // Specify your CorDapp's dependencies below, including dependent CorDapps.
     // We've defined Cash as a dependent CorDapp as an example.
-    cordapp "net.corda:corda-finance:$corda_release_version"
     cordapp project(":states-and-contracts")
-    cordapp project(":other-sources")
 }
 
 task integrationTest(type: Test, dependsOn: []) {
@@ -92,52 +89,4 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         jvmTarget = "1.8"
         javaParameters = true   // Useful for reflection.
     }
-}
-
-task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
-    directory "./build/nodes"
-    networkMap "O=Controller,L=London,C=GB"
-    node {
-        name "O=Controller,L=London,C=GB"
-        advertisedServices = ["corda.notary.validating"]
-        p2pPort 10002
-        rpcPort 10003
-        cordapps = [
-                "com.template:states-and-contracts:$version",
-                "com.template:other-sources:$version",
-                "net.corda:corda-finance:$corda_release_version"
-        ]
-    }
-    node {
-        name "O=PartyA,L=London,C=GB"
-        advertisedServices = []
-        p2pPort 10005
-        rpcPort 10006
-        webPort 10007
-        cordapps = [
-                "com.template:states-and-contracts:$version",
-                "com.template:other-sources:$version",
-                "net.corda:corda-finance:$corda_release_version"
-        ]
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": []]]
-    }
-    node {
-        name "O=PartyB,L=New York,C=US"
-        advertisedServices = []
-        p2pPort 10008
-        rpcPort 10009
-        webPort 10010
-        cordapps = [
-                "com.template:states-and-contracts:$version",
-                "com.template:other-sources:$version",
-                "net.corda:corda-finance:$corda_release_version"
-        ]
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": []]]
-    }
-}
-
-task runTemplateClient(type: JavaExec) {
-    classpath = sourceSets.main.runtimeClasspath
-    main = 'com.template.ClientKt'
-    args 'localhost:10006'
 }
