+++ other-sources/src/main/kotlin/com/template/App.kt
@@ -31,32 +31,6 @@ class TemplateApi(val rpcOps: CordaRPCOps) {
     }
 }
 
-// *****************
-// * Contract Code *
-// *****************
-// This is used to identify our contract when building a transaction
-val TEMPLATE_CONTRACT_ID = "com.template.TemplateContract"
-
-open class TemplateContract : Contract {
-    // A transaction is considered valid if the verify() function of the contract of each of the transaction's input
-    // and output states does not throw an exception.
-    override fun verify(tx: LedgerTransaction) {
-        // Verification logic goes here.
-    }
-
-    // Used to indicate the transaction's intent.
-    interface Commands : CommandData {
-        class Action : Commands
-    }
-}
-
-// *********
-// * State *
-// *********
-data class TemplateState(val data: String) : ContractState {
-    override val participants: List<AbstractParty> get() = listOf()
-}
-
 // *********
 // * Flows *
 // *********
