+++ other-sources/src/main/resources/META-INF/services/net.corda.core.serialization.SerializationWhitelist
