+++ other-sources/src/main/resources/META-INF/services/net.corda.webserver.services.WebServerPluginRegistry
