+++ other-sources/src/test/kotlin/com/template/ContractTests.kt
