+++ settings.gradle
@@ -1,0 +1,2 @@
+include 'states-and-contracts'
+include 'other-sources'
